added user auth endpoints

+++ Controllers/RecipeController.cs
@@ -1,6 +1,9 @@
-using Microsoft.AspNetCore.Mvc;
-using feastly_api.Repositories;
 using feastly_api.Models;
+using feastly_api.Repositories;
+using Microsoft.AspNetCore.Mvc;
+
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 
 namespace feastly_api.Controllers
 {
@@ -15,6 +18,16 @@ namespace feastly_api.Controllers
         {
             _logger = logger;
             _recipeRepository = repository;
+        }
+
+        private int? GetUserIdFromToken()
+        {
+            var userIdClaim = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "UserId");
+            if (userIdClaim == null || !int.TryParse(userIdClaim.Value, out var userId))
+            {
+                return null;
+            }
+            return userId;
         }
 
         [HttpGet]
@@ -37,6 +50,7 @@ namespace feastly_api.Controllers
             return Ok(recipe);
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpPost]
         public async Task<ActionResult<Recipe>> CreateRecipe([FromBody] Recipe newRecipe)
         {
@@ -45,11 +59,19 @@ namespace feastly_api.Controllers
                 return BadRequest();
             }
 
+            var userId = GetUserIdFromToken();
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
+            newRecipe.UserId = userId.Value;
+
             var createdRecipe = await _recipeRepository.CreateRecipe(newRecipe);
             return CreatedAtAction(nameof(GetAllRecipes), new { recipeId = createdRecipe.RecipeId }, createdRecipe);
         }
 
-
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpPut]
         [Route("{recipeId:int}")]
         public async Task<ActionResult<Recipe>> UpdateRecipe([FromBody] Recipe updatedRecipe, int recipeId)
@@ -59,6 +81,14 @@ namespace feastly_api.Controllers
                 return BadRequest();
             }
 
+            var userId = GetUserIdFromToken();
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
+            updatedRecipe.UserId = userId.Value;
+
             var recipeUpdate = await _recipeRepository.UpdateRecipe(updatedRecipe);
             if (recipeUpdate == null)
             {
@@ -67,14 +97,28 @@ namespace feastly_api.Controllers
             return Ok(recipeUpdate);
         }
 
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         [HttpDelete("{recipeId:int}")]
         public async Task<IActionResult> DeleteRecipe(int recipeId)
         {
+            var userId = GetUserIdFromToken();
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
             var existingRecipe = await _recipeRepository.GetRecipe(recipeId);
+          
             if (existingRecipe == null)
             {
                 return NotFound();
             }
+
+            if (existingRecipe.UserId != userId)
+            {
+                return Forbid();
+            }
+            
             await _recipeRepository.DeleteRecipe(recipeId);
             return NoContent();
         }
